Update V0.04 - QOL Fix
- Unless you have Debug mode on (the in game option) you can no longer spawn an NPC while pressing 6 key. [THANKS TO NIGHTSCALE FOR FINDING THIS]
- Debug mode in the options actually will turn on the debug features.
- Marked off some of the "print()" functions, it may not seem like much, but enough of them running at once will start adding lag. So, removed some leftovers from the previous modders.

- Carefully updated Task_IsPursue_SC() function. Now it will let any NPC scan for humans, but as long as it doesn't have a target and along with several other if's.
Basically what this means is, the NPCs will have a small radius sense around them to know when there's another hostile around them. (does it always work? Mostly)
But generally this is meant for hostile NPCs the most. This way if a NPC gets behind another while an attack happens, the npc can turn around. to simulate that they know that there was just previously a fight going on, and then continue the fight.