Update V0.05 - Follow Task Tweaks
First get go at trying to make the follow task work.

Also there is an update being worked on in the options to change Alt spawning further. Instead of making a group with %, it will now be separated. the context menu text isn't filled though.
I just want to update this progress for anyone wanting to test. I've been awake for 40 hours working on this with 2 hours of sleep.
https://i.imgur.com/wRXLlDI.png