BackupUpdate (Unknown-stable)
This phase I do not recommend trying yet, I am saving this for now, because I just got the NPCs to stop getting stuck in front of *any* door (a but I created during pursue task) and I believe I fixed it. But I want to backup progress before continuing on.